Breaking out the environment class to its own file
